backend/requirements.txt
flask
flask-cors
firebase-admin
gunicorn
google-cloud-storage